fully added support for linux server client + many fixes + time to test

+++ QVPN/QVPN Lib/include/Linux/qvpn_net_tools.hpp
@@ -11,7 +11,7 @@
 
 #include <utility>
 
-using SOCKET = int;
+using SOCKET = long int;
 
 
 namespace std
